requirements/base.txt
asgiref==3.3.4
blinker==1.4
certifi==2020.12.5
click==7.1.2
Django==3.2.1
django-filter==2.4.0
django-framework==0.1.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.4
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mongoengine==0.23.1
pymongo==3.11.4
pytz==2021.1
sentry-sdk==0.6.5
sqlparse==0.4.1
urllib3==1.26.4
Werkzeug==1.0.1
